+++ ntex-io/src/filter.rs
@@ -94,7 +94,7 @@ impl Filter for Base {
     }
 
     fn poll_write_ready(&self, cx: &mut Context<'_>) -> Poll<Readiness> {
-        if self.0.0.flags.is_terminated() {
+        if self.0.0.flags.is_terminated() || self.0.0.flags.is_terminating() {
             Poll::Ready(Readiness::Terminate)
         } else {
             self.0.0.write_task.register(cx.waker());