+++ ntex-net/src/compio/io.rs
@@ -122,7 +122,6 @@ where
     if !ctx.is_stopped() {
         let result = poll_fn(|cx| ctx.shutdown(true, cx)).await;
         log::trace!("{}: Shuting down complete {result:?}", ctx.tag());
-        ctx.stop(None);
     }
 }
 
@@ -164,7 +163,8 @@ where
                         break;
                     }
                 } else if write_buf(&mut io, ctx, bufs).await == IoTaskStatus::Stop {
-                    let _ = io.shutdown().await;
+                    let res = io.shutdown().await;
+                    ctx.stopped(res.err());
                     break;
                 }
             }
@@ -172,10 +172,14 @@ where
                 let bufs = ctx.with_write_buf(build_bufs);
                 write_buf(&mut io, ctx, bufs).await;
                 let res = io.shutdown().await;
-                ctx.stop(res.err());
+                ctx.stopped(res.err());
                 break;
             }
-            Readiness::Terminate => return,
+            Readiness::Terminate => {
+                let res = io.shutdown().await;
+                ctx.stopped(res.err());
+                return;
+            }
         }
     }
 }