+++ ntex-net/src/polling/io.rs
@@ -109,5 +109,5 @@ async fn run(ctl: StreamCtl, context: ntex_io::IoContext) {
 
     let result = ctl.shutdown().await;
     log::trace!("{}: Shutdown complete {result:?}", context.tag());
-    context.stop(result.err());
+    context.stopped(result.err());
 }